OPSMN-3649: Marked the object that tracks modified attributes as serializable.
When bulk importing consents, the importer reads all records from input file
and builds one consent response object per participant with all participant
responses. Since maintaining all records in-memory can be taxing, the excess
records are spooled to disk using object serialisation technique. However
during this serialisation process, the list of object attributes that got
modified, which is used in implementing patch update, is lost. As a result,
when the serialised object is fetched back from disk, the modified attributes
list is empty, which causes the patch update API to believe there is nothing
to update. This behaviour can be observed only if input file has consents for
more than 100 participants.

Marked the object that tracks modified attributes as Serializable so that when
the object is written to disk, the modified attributes list is also persisted.