OPSMN-3483: Add null guard when an invalid specimen lineage is specified
An invalid lineage results in an error and the lineage field is not set.
However, the specimen builder proceeds with other steps of object
creation/assembly to find out more errors, if any. During this process,
the builder encounters null lineage, which was causing NPE. To prevent
NPE, added null guards for lineage.