OPSMN-3591: Initialise user as regular type during user object assembly process.
When an invalid user type is specified, user type is set to null. This causes
other steps in assembly process that depends on user type to fail with NPE.
Therefore always initialise user to regular type, which will remain as is or
over written by input type. This ensures other steps that depend on user
type do not fail with NPE.