OPSMN-3743: Remember the requested page when landing page is rendered.
Earlier all landing pages were children of a common unsigned state.
This unsigned state was used to determine whether to remember the
requested page or not.

However, specimen catalog dashboard is not direct child of common
unsigned state. Therefore when it is used as landing page, the requested
page is not remembered and therefore sign-in always navigates users to
CP list view.

Fix: Explcitly marked the catalog dashboard as landing state so that
the requested page can be remembered for navigating user to on successful
sign-in.