OPSMN-3747: Ensured correct column metadata is used to format column values.
Cause:
When results are screened, column n data is formatted using metadata of
column (n - 1). This difference occurs because, for screening purpose,
CP ID column is automatically added to determine whether the row PHI
fields should be masked or not. This results in offsetting data value
positions by 1 in row array.