OPSMN-3753: Save specimen position objects directly to DB.
For bounded containers, specimen position ordinals are fetched into
memory for computing next available position in the container. This
works fine as the no. of slots to examine are limited (81 for 9 x 9,
100 for 10 x 10 box). However, the same logic was executed for
dimensionless containers which do not have the concept of any layout
or position coordinates. As a result, for large dimensionless
containers, many slots were examined before storing a new specimen
object.

The fix is to not examine any slots of dimensionless container and
store the specimen position directly to DB.