OPSMN-3808: JavaScript implementation of Date object/type is broken.
JavaScript specs and implementation do not take into account the
discontinuity of time at various moments in history of different
geographic locations because moving time either forward or backward.

Fix: As a stop gap arrangement, fixed only participant date of birth,
which can go back as far 19th century.

The UI sends date of birth in yyyy-MM-dd format. The server parses
this date in local timezone and stores in database.

The server API sends both the UTC time (no. of milliseconds elapsed
since Epoch) and the local time in yyyy-MM-dd format. The UI uses
the local string representation for display to end users.

A more thorough treatment to this problem is beyond the scope of
v4.1