OPSMN-3849: Deserialise input custom fields data into extension detail interaction object.
During implementation of export custom fields, a new method getAttrsMap() was added
to retrieve custom fields data to be exported to CSV file. However, this was not
needed in API response. Therefore the method was appropriately marked with
JsonIgnore annotation. However, Jackson interpreted that both deserialisation
and serialisation are not needed. As a result, the custom fields data
passed in API request was not deserialised into interaction object.

Instructed Jackson not to serialise but deserialise the custom fields data
by using @JsonProperty on setAttrsMap() method.