OPSMN-3061: DE library with below fix:
OPSMN-3061: Added activity condition as part of join condition so that
the records of parent entity are not filtered because of absence of child
entity records.

for example (p is parent table, c is child table):

select
  p.*, c.*
from
  p left join c on c.p_id = p.id
where
  p.status = 'Active' and
  c.status = 'Active'

is transformed to

select
  p.*, c.*
from
  p left join c on c.p_id = p.id and c.status = 'Active'
where
  p.status = 'Active'

This ensure p records are still available for selection in output
even if c records are absent.