OPSMN-3863: Fixed the query and DAO API to obtain DP list based on input criteria.
For non-admin users, there was a join between DP and sites tables for access
checks. This created one DP record in result set for each site of institute on
which the user has DP read rights. Further, the result set was restricted to 100 records.
Therefore when a user with DP read rights on all sites of an institute invoked the query,
the same DP appeared in all 100 records of the result set. The DAO API subsequently
de-duplicated the records resulting in a single DP record being returned to the user.

Fix:
1. Used sub-query to obtain list of distinct DP IDs on which the user has
   read rights
2. The results of above sub-query is used to fetch first hundred DP records.
3. Gets rid of in-memory de-duplication.

select
  dp
from
  dps
where
 dp.id in (
   select
     distinct accessible_dp_ids
   from
     <table_joins>
   where
     <input_criteria>
 )
order by
  dp.title
limit 0, 100