OPSMN-3892: Added correct pre-condition to drop FK only if it exists in database.
The FK on address ID fields of sites and users table needs to be dropped.
However it has been observed that some databases have FK on address ID field
and some others don't have. Therefore a pre-condition was added to drop
FK only if present. Unfortunately, the pre-condition was probing for
a different FK i.e. if (abc present) drop xyz

Fix: Ensured the pre-condition probes for the presence of FK that is being
dropped i.e if (abc present) drop abc