OPSMN-3898: Configurable participant lookup and add workflow.
Participant registration is broken into 3 steps:

Lookup -> Select Match -> Review & Register

The lookup step is displayed based on the lookup fields configured
either at the CP or system level. If no lookup fields are configured
then all the participant fields configured in the dictionary are
used to display the lookup step. Lookup fields are those that are
marked in dictionary with attribute participantLookup: true

The select match step shows matching participants table. The table
columns are made of participant fields configured in the dictionary

The review and register step opens up the select matched in edit
mode for review and submitting the changes, if any.

Cases:
If there are no matching participants, the select match step is
skipped and whatever user entered in lookup step appears in
review and register step.

If there is a single match based on unique attribute like eMPI, the
match is auto selected and the select match step is skipped.

If no lookup fields are configured, and there are no matching
participants, then the participant is registered based on the
details sepecified by user (last 2 steps are skipped).