OPSMN-3954: Used a light purple colour to display count of unstored collected specimens.
Earlier the colour used for virtual specimens was Orange which is used for pending
specimens too. This confused users when they looked at vist stats graph and what they
saw in the visit specimen tree.

Used light purple color for virtual specimens in visit stats graph and in the specimen
tree.