OPSMN-3579: The file download URL is stored in a shadow variable $$fileDownloadUrl. This is done so that the JSON submitted to server is free from the URL field. However, the HTML minifier removes $ from the variable name when used in interpolation string resulting in following transformation:
From:  <a href="{{attr.$$fileDownloadUrl}}"> ... </a>
To  : <a href="{{attr.$fileDownloadUrl}}"> ... </a>

Above is an invalid transformation, as attr.$fileDownloadUrl is a non-existing property
resulting in an empty link. The end result is users are unable to download the file.

To fix the above issue, used a local scope variable that is initialised to file download
URL value to interpolate the href attribute value of anchor tag <a>