OPSMN-4067: Fixed the OS date picker widget to reject all date strings that fail to conform system level date format settings.
The problem was the Angular bootstrap date widget was using "MM/dd/yyyy"
incremental format as the fallback when the input string didn't conform to
system level date format. This made users believe the date picker has
gone crazy and accepts all sorts of things.

Added a model parser to reject all input strings that are not in
system level date format.